+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 - Added feature toggles !2633 @martincodes-de
 - The backend now provides the map tile api key for the client #1074 !2621 !2925 @alex.simm
 - Bump vue-advanced-chat to 2.0.10 !2927 @chriswalg
+- Added a rate limiter to the login endpoint !2672 @alex.simm
 
 ## Development documentation